Apply template

+++ cloud_app/views.py
@@ -18,29 +18,11 @@ from django.shortcuts import render, redirect, get_object_or_404
 
 # from cloud.forms import FileUploadForm, UserForm, FileEditForm
 from django.template.defaulttags import comment
+from django.utils.decorators import method_decorator
+from django.views.decorators.csrf import csrf_exempt
 
 from .forms import UserForm, FileUploadForm, UserEditForm, FileEditForm, CommentForm
 from .models import FileDetailInfo, Comment, UserInfo
-
-
-# 본인 업로드, 공유받은 파일 리스트 출력
-# def main(request):
-#     files = FileDetailInfo.objects.order_by('-file_upload')
-#     context = {'files': files}
-#     return render(request, 'cloud/main.html', context)
-
-
-#회원정보수정, UserForm(휴대전화번호) 변경 저장 후 메인페이지 이동
-# def user_edit(request, pk):
-#     user = get_object_or_404(UserInfo, pk=UserInfo.id)
-#     if request.method == "POST":
-#         form = UserEditForm(request.POST, instance=user)
-#         if form.is_valid():
-#             user = form.save()
-#             return redirect('main')
-#     else:
-#         form = UserEditForm()
-#     return render(request, 'cloud/user_edit.html', {'form': form})
 
 
 def user_edit(request):
@@ -53,26 +35,100 @@ def user_edit(request):
         form = UserEditForm(request.user)
     return render(request, 'cloud/user_edit.html', {'form': form})
 
-
-#파일 업로드, FileUploadForm(파일명, 파일업로드, 게스트, 게시일자)출력 저장 후 메인 or 해당 파일상세페이지 이동
 def file_upload(request):
-    if request.method == "POST":
-        form = FileUploadForm(request.POST)
-        if form.is_valid():
-            file = form.save(commit=False)
-            file.file_upload = timezone.now()
-            file.save()
-            return redirect('main')
+    user = UserInfo.objects.filter(name=request.session.get('name')).get()
+    if request.method == 'POST':
+        form = FileUploadForm(request.POST, request.FILES, request.session)
+        print('aaaaaaa')
+        print('id')
+        print('bbbbbbb')
+        id = 1
+        files = FileDetailInfo.objects.all()
+        maxid = 0
+        for file in files:
+            maxid = max(maxid, file.id)
+        id = maxid + 1
+        print(id)
+        file_title = request.POST['file_title']
+        guest_name = request.POST['guest_name']
+        file_url = request.FILES.get('file_url')
+        owner_name = user
+        # file_url = request.FILES['file_url']
+        print('-----------')
+        print(request.FILES)
+        print('-----------')
+        file = FileDetailInfo(
+            id=id,
+            file_title=file_title,
+            guest_name=guest_name,
+            file_url=file_url,
+            owner_name=owner_name,
+        )
+        file.save()
+        return redirect('main')
     else:
         form = FileUploadForm()
-    return render(request, 'cloud/file_upload.html', {'form': form})
+    return render(request, 'cloud/file_upload.html', {'form': form, 'userinfo': user, })
+# def file_upload(request):
+#     user = UserInfo.objects.filter(name=request.session.get('name')).get()
+#     if request.method == 'POST':
+#         form = FileUploadForm(request.POST, request.FILES, request.session)
+#         print('aaaaaaa')
+#         print('id')
+#         print('bbbbbbb')
+#         id=1
+#         for x in range(1, 100):
+#             id+= 1
+#             id.insert_one({"id": id})
+#         file_title = request.POST['file_title']
+#         guest_name = request.POST['guest_name']
+#         file_url = request.FILES.get('file_url')
+#         owner_name = user
+#         # file_url = request.FILES['file_url']
+#         print('-----------')
+#         print(request.FILES)
+#         print('-----------')
+#         file = FileDetailInfo(
+#             id=id,
+#             file_title=file_title,
+#             guest_name=guest_name,
+#             file_url=file_url,
+#             owner_name=owner_name,
+#         )
+#         file.save()
+#         return redirect('main')
+#     else:
+#         form = FileUploadForm()
+#     return render(request, 'cloud/file_upload.html', {'form': form, 'userinfo': user, })
 
 
 def file_detail(request, pk):
+    user = UserInfo.objects.filter(name=request.session.get('name')).get()
     file = get_object_or_404(FileDetailInfo, pk=pk)
-    comments = Comment.objects.filter(file_detail = file.pk)
+    print(user)
+    print(file.owner_name)
+    guests = file.guest_name.strip().split(",")
+    comments = Comment.objects.filter(file_detail=pk)
     comment_form = CommentForm()
-    return render(request, 'cloud/file_detail.html', {'file': file, 'comments':comments, 'comment_form': comment_form})
+    if request.method == "POST":
+        file_detail = file
+        comment_text = request.POST['body']
+        comment_name = user
+        comment = Comment(
+            file_detail=file_detail,
+            comment_text=comment_text,
+            comment_name=comment_name,
+        )
+        comment.save()
+        return redirect('file_detail', pk)
+    # form = CommentForm(request.POST)
+    # if form.is_valid():
+    #     form = form.save(commit=False)
+    #     comment.comment_name = user
+    #     form.save()
+    #     return redirect('file_detail', pk)
+    context = {'file': file, 'comments':comments,'comment_form':comment_form, 'userinfo':user, 'guests':guests}
+    return render(request, 'cloud/file_detail.html', context)
 
 
 def sign_up(request):
@@ -90,30 +146,71 @@ def sign_up(request):
 
 
 def file_edit(request, pk):
+    user = UserInfo.objects.filter(name=request.session.get('name')).get()
     file = get_object_or_404(FileDetailInfo, pk=pk)
+    if user != file.owner_name:
+        messages.error(request, '수정할 수 없습니다')
+        return redirect('file_detail', pk=file.id)
     if request.method == "POST":
-        form = FileEditForm(request.POST, instance=file)
-        if form.is_valid():
-            file = form.save()
-            return redirect('file_detail', pk=file.pk)
+        form = FileUploadForm(request.POST, request.FILES, request.session)
+        # if form.is_valid():
+        #     file = form.save()
+        #     return redirect('file_detail', pk=file.pk)
+        form = FileEditForm(request.POST, request.session)
+        file_title = request.POST['file_title']
+        guest_name = request.POST['guest_name']
+        file_url = request.FILES.get('file_url')
+        owner_name = user
+        id=file.id
+        file = FileDetailInfo(
+            id=id,
+            file_title=file_title,
+            guest_name=guest_name,
+            file_url=file_url,
+            owner_name=owner_name,
+        )
+        file.save()
+        file_pk = pk
+        return redirect('file_detail', file_pk)
     else:
-        form = FileEditForm(instance=file)
-    return render(request, 'cloud/file_edit.html', {'form': form})
+        form = FileUploadForm(instance=file)
+    return render(request, 'cloud/file_edit.html', {'form': form, 'userinfo': user, })
 
 
 def file_remove(request, pk):
+    user = UserInfo.objects.filter(name=request.session.get('name')).get()
     file = get_object_or_404(FileDetailInfo, pk=pk)
+    if user != file.owner_name:
+        messages.error(request, '삭제할 수 없습니다')
+        return redirect('file_detail', pk=file.id)
     file.delete()
     return redirect('main')
 
-
-def add_comment(request, pk):
-    form = CommentForm(request.POST)
-    if form.is_valid():
-        form = form.save(commit=False)
-        comment.comment_name = request.user
-        form.save()
-        return redirect('file_detail', pk)
+# @method_decorator(csrf_exempt)
+# def add_comment(request, pk):
+#     user = UserInfo.objects.filter(name=request.session.get('name')).get()
+#     # file = fileDetailInfo.objects.filter(file.id=pk)
+#     if request.method == 'POST':
+#         form = CommentForm(request.POST, request.session)
+#         comment_text = request.POST.get('comment_text')
+#         file_detail = file
+#         comment_name = user
+#         comment = Comment(
+#             file_detail=file_detail,
+#             comment_text=comment_text,
+#             comment_name=comment_name,
+#         )
+#         comment.save()
+#         file_pk = comment.file_details
+#         return redirect('file_detail', pk=file_pk)
+# def add_comment(request, pk):
+#     user = UserInfo.objects.filter(name=request.session.get('name')).get()
+#     form = CommentForm(request.POST)
+#     if form.is_valid():
+#         form = form.save(commit=False)
+#         comment.comment_name = user
+#         form.save()
+#         return redirect('file_detail', pk)
 
 
 def remove_comment(request, pk):
@@ -123,15 +220,36 @@ def remove_comment(request, pk):
     return redirect('file_detail', pk=file_pk)
 
 
-def comment_edit(request):
-    jsonObject = json.loads(request.body)
-    comment = Comment.objects.filter(id=jsonObject.get('id'))
-    context = {'result':'no'}
-    if comment is not None:
-        comment.update(content=jsonObject.get('content'))
-        context = {'result':'ok'}
-        return JsonResponse(context);
-    return JsonResponse(context)
+def comment_edit(request, pk):
+    user = UserInfo.objects.filter(name=request.session.get('name')).get()
+    comment = get_object_or_404(Comment, pk=pk)
+    if request.method == "POST":
+        form = CommentForm(request.POST, request.session)
+        comment_text = request.POST['comment_text']
+        comment_name = user
+        file_detail = comment.id
+        comment = CommentForm(
+            comment_text = comment_text,
+            comment_name = comment_name,
+            file_detail = file_detail,
+        )
+        comment.save()
+        return redirect('file_detail', pk=comment.file_detail)
+    else:
+        form = CommentForm(instance=comment)
+    context = {'comment':comment, 'form':form}
+    return render(request, 'cloud/comment_edit.html', context)
+
+
+# def comment_edit(request):
+#     jsonObject = json.loads(request.body)
+#     comment = Comment.objects.filter(id=jsonObject.get('id'))
+#     context = {'result':'no'}
+#     if comment is not None:
+#         comment.update(content=jsonObject.get('content'))
+#         context = {'result':'ok'}
+#         return JsonResponse(context);
+#     return JsonResponse(context)
 
 
 def login(request: HttpRequest) -> HttpResponse:
@@ -175,8 +293,20 @@ def main(request: HttpRequest):
         print(user.name)
         print("logged in")
         files = FileDetailInfo.objects.filter(owner_name=user.email).order_by('-file_upload')
-        print(files)
-        context = {'files': files, 'userinfo': user}
+        print("files >> ", files)
+        all_files = FileDetailInfo.objects.all()
+        print("all_files >> ", all_files)
+        shared_files = []
+        for file in all_files:
+            guests = file.guest_name.strip().split(",")
+            for guest in guests:
+                guest = guest.strip()
+                if name == guest:
+                    shared_files.append(file)
+                elif user.email == guest:
+                    shared_files.append(file)
+        print("shared_files >> ", shared_files)
+        context = {'files': files, 'userinfo': user, 'shared_files': shared_files}
         return render(request, 'cloud/main.html', context)
     else:
         print("not logged in")
